% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/america.R
\name{america}
\alias{america}
\title{Derivati americani}
\usage{
america(X, assets, errore = 8, all.ports = FALSE)
}
\arguments{
\item{X}{Processo stocastico del payoff del derivato americano.}

\item{assets}{Lista contenente i processi stocastici dei prezzi dei titoli considerati.}

\item{errore}{Numero di decimali da considerare per l'approssimazione per vmin.}

\item{all_ports}{Paramentro per scegliere se costruire a prescindere tutti i portafogli o solo quelli razionali.}
}
\value{
Lista contenente i prezzi di non arbitraggio, i portafogli di replica e le strategie di arresto vmin e vmax (se questi esistono).
}
\description{
Funzione che calcola i prezzi di non arbitraggio, le repliche e le strategie di arresto di un derivato americano.
Si assume che il numeraire (S0) sia un processo deterministico e che il mercato sia completo.
}
\examples{
assets = binomiale(S1 = 1, r = 0, u = 2, d = 1/2, N = 3)
X = list(2/3, c(5/3, 1/6), c(0, 2/3, 2/3, 0), c(0, 0, 5/3, 1/6, 5/3, 1/6, 1/6, 0))
america(X = X, assets = assets, errore = 8, all.ports = FALSE)
}
