% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomiale.R
\name{binomiale}
\alias{binomiale}
\title{Mercato binomiale}
\usage{
binomiale(S1, r, u, d, N)
}
\arguments{
\item{S1}{Valore al tempo 0 del titolo rischioso.}

\item{r}{Tasso di interesse periodale.}

\item{u}{Uptick.}

\item{d}{Downtick.}

\item{N}{Numero di periodi considerati per la filtrazione.}
}
\value{
Lista contenente il mercato binomiale.
}
\description{
Funzione che prepara un mercato binomiale con un titolo rischioso.
}
\examples{
binomiale(S1 = 1, r = 1/2, u = 2, d = 1/2, N = 3)
}
