% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doob.R
\name{doob}
\alias{doob}
\title{Decomposizione di Doob}
\usage{
doob(X, P)
}
\arguments{
\item{X}{Processo stocastico da decomporre.}

\item{P}{Misura di probabilità usata per la decomposizione.}
}
\value{
Lista contenente le componenti Y e Z della decomposizione di Doob.
}
\description{
Funzione che calcola la decomposizione di Doob di un processo stocastico X data una misura di probabilità P
}
\examples{
X = list(3, c(2, 2, 3), c(1, 2, 3, 4, 5, 6, 7))
X = filtrazione(list(X), blocchi = list(c(3, 2, 2)))[[1]]
P = c(1/4, 0, 1/4, 0, 1/4, 0, 1/4)
doob(X = X, P = P)
}
