% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/europa.R
\name{europa}
\alias{europa}
\title{Derivati europei}
\usage{
europa(X, assets)
}
\arguments{
\item{X}{Variabile casuale del payoff del derivato europeo.}

\item{assets}{Lista contenente i processi stocastici dei prezzi dei titoli considerati.}
}
\value{
Lista contenente i prezzi di non arbitraggio ed i portafogli di replica (se esistono).
}
\description{
Funzione che calcola lea repliche ed i prezzi di non arbitraggio di un derivato europeo.
}
\examples{
#Esempio 1: Esiste ed è unica la misura martingala
S0 = list(2, c(4, 1), c(6, 2, 3, 1), c(8, 4, 6, 1, 6, 2, 4, 1/2))
S1 = list(4, c(8, 3), c(12, 2, 9, 4), c(16, 4, 12, 1, 12, 8, 8, 2))
X = c(8, 0, 22/3, 0, 20/3, 8/3, 13/3, 0)
assets = filtrazione(list("S0" = S0, "S1" = S1), blocchi = list(c(2,2), c(2,2,2,2)))
europa(X = X, assets = assets)

#Esempio 2: Non esiste una misura martingala
S0 = list(1, c(1, 1), c(1, 1, 1, 1, 1))
S1 = list(3, c(2, 7), c(1, 3, 2, 2, 6))
S2 = list(3/2, c(1, 2), c(1, 2, 0, 1, 3))
X = c(0, 1, 2, 1, 3)
assets = filtrazione(list("S0" = S0, "S1" = S1, "S2" = S2), list(2, c(3, 2)))
europa(X = X, assets = assets)
}
