% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtrazione.R
\name{filtrazione}
\alias{filtrazione}
\title{Preparazione della filtrazione}
\usage{
filtrazione(assets, blocchi)
}
\arguments{
\item{assets}{Lista contenente i processi stocastici considerati.}

\item{blocchi}{Lista che, per ogni tempo da 1 a N-1, contiene il numero di blocchi della partizione successiva che nascono da quel blocco.}
}
\value{
Lista contenente la filtrazione.
}
\description{
Funzione che prepara la filtrazione alle funzioni di stocastico.
Con iBj si indica il j-esimo blocco di una partizione che nasce dal blocco numero i della partizione precedente.
}
\examples{
S0 = list(1, c(1,2), c(1,2,3,4,5), c(1,2,3,4,5,6,7,8,9, 10, 11, 12))
S1 = list(2, c(3,4), c(5,6,7,8,9), c(10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20))
assets = list("S0" = S0, "S1" = S1)
blocchi = list(c(3, 2), c(2, 3, 2, 3, 2))
filtrazione(assets = assets, blocchi = blocchi)
}
