% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ito.R
\name{Ito}
\alias{Ito}
\title{Lemma di Ito}
\usage{
Ito(Yt = "Xt", dXt = "0*dt + 1*dBt")
}
\arguments{
\item{Yt}{Funzione che determina il processo stocastico rispetto a cui si applica il lemma.}

\item{dXt}{Equazione differenziale associata al processo Xt di partenza.}
}
\value{
Equazione differenziale stocastica di cui Yt è soluzione.
}
\description{
Funzione che determina l'equazione differenziale stocastica associata ad un processo stocastico generico Xt utilizzando il Lemma di Ito.
}
\examples{
#Esempio 1: Moto Browniano
Yt = "mu*t + sigma*Xt"
dXt = "0*dt + 1*dBt"
Ito(Yt = Yt, dXt = dXt)

#Esempio 2: Moto Browniano geometrico
Yt = "X0*exp(mu*t + sigma*Xt)"
dXt = "0*dt + 1*dBt"
Ito(Yt = Yt, dXt = dXt)

#Esempio 3
Yt = "log(Xt)"
dXt = "mu*Xt*dt + sigma*Xt*dBt"
Ito(Yt = Yt, dXt = dXt)
}
