% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin.sys.R
\name{lin.sys}
\alias{lin.sys}
\title{Sistemi lineari}
\usage{
lin.sys(A, b, nomi = NA)
}
\arguments{
\item{A}{Matrice dei coefficienti del sistema (dimensione n x p).}

\item{b}{Vettore dei termini noti (vettore di lunghezza n).}

\item{nomi}{Nomi da associare ai risultati (vettore di lunghezza p).}
}
\value{
Soluzione del sistema.
}
\description{
Funzione che risolve sistemi lineari.
}
\examples{
#Esempio 1: Soluzione unica
A = matrix(c(1, 2, 3, 8, 2, 1, 1, 6, 4), ncol = 3)
b = c(3, 1, 5)
lin.sys(A = A, b = b)

#Esempio 2: Infinite soluzioni
A = matrix(c(1, 1, 1, 1, 3/2, 2), ncol = 3, byrow = TRUE)
b = c(1, 3/2)
lin.sys(A = A, b = b)

#Esempio 3: Nessuna soluzione
A = matrix(c(1, 2, 3, 4, 5, 6, 7, 8, 9), ncol = 3)
b = c(3, 1, 5)
lin.sys(A = A, b = b)
}
