% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/martingala.R
\name{martingala}
\alias{martingala}
\title{Calcolo di misura martingala}
\usage{
martingala(assets)
}
\arguments{
\item{assets}{Lista contenente i processi stocastici dei prezzi dei titoli considerati.}
}
\value{
Lista contenente la misura martingala (se esiste) e le probabilità condizionate dei blocchi.
}
\description{
Funzione che calcola la misura martingala di un mercato.
}
\examples{
#Esempio 1: Esiste ed è unica la misura martingala
S0 = list(2, c(4, 1), c(6, 2, 3, 1), c(8, 4, 6, 1, 6, 2, 4, 1/2))
S1 = list(4, c(8, 3), c(12, 2, 9, 4), c(16, 4, 12, 1, 12, 8, 8, 2))
assets = filtrazione(list("S0" = S0, "S1" = S1), blocchi = list(c(2, 2), c(2, 2, 2, 2)))
martingala(assets = assets)

#Esempio 2: Esiste più di una misura martingala
S0 = list(1, c(1, 1), c(1, 1, 1, 1, 1))
S1 = list(2, c(3/2, 4), c(1, 3/2, 2, 2, 4))
assets = filtrazione(list("S0" = S0, "S1" = S1), blocchi = list(c(3, 2)))
martingala(assets = assets)

#Esempio 3: Non esiste una misura martingala
S0 = list(1, c(1, 1), c(1, 1, 1, 1, 1))
S1 = list(3, c(2, 7), c(1, 3, 2, 2, 6))
S2 = list(3/2, c(1, 2), c(1, 2, 0, 1, 3))
assets = filtrazione(list("S0" = S0, "S1" = S1, "S2" = S2), list(2, c(3, 2)))
martingala(assets = assets)
}
