% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_filtrazione.R
\name{plot_filtrazione}
\alias{plot_filtrazione}
\title{Rappresentazione grafica della filtrazione}
\usage{
plot_filtrazione(assets, col1 = "lightblue", col2 = "red")
}
\arguments{
\item{assets}{Lista contenente i processi stocastici considerati.}

\item{col1}{Colore assegnato ai rettangoli contenenti i blocchi della partizione.}

\item{col2}{Colore assegnato al testo dei valori dei processi stocastici.}
}
\value{
Grafico della filtrazione.
}
\description{
Funzione che disegna il grafico della filtrazione inserita con i relativi valori dei processi stocastici.
}
\examples{
S0 = list(1, c(1,2), c(1,2,3,4,5))
S1 = list(2, c(3,4), c(5,6,7,8,9))
assets = list("S0" = S0, "S1" = S1)
blocchi = list(c(3, 2), c(2, 3, 2, 3, 2))
assets = filtrazione(assets = assets, blocchi = blocchi)
plot_filtrazione(assets = assets, col1 = "lightblue", col2 = "red")
}
