% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trinomiale.R
\name{trinomiale}
\alias{trinomiale}
\title{Mercato trinomiale}
\usage{
trinomiale(S1, S2, r, u, m, d, N)
}
\arguments{
\item{S1}{Valore al tempo 0 del titolo rischioso 1.}

\item{S2}{Valore al tempo 0 del titolo rischioso 2.}

\item{r}{Tasso di interesse periodale.}

\item{u}{Vettore degli uptick dei due titoli.}

\item{m}{Vettore dei mediumtick dei due titoli.}

\item{d}{Vettore dei downtick dei due titoli.}

\item{N}{Numero di periodi considerati per la filtrazione.}
}
\value{
Lista contenente il mercato trinomiale.
}
\description{
Funzione che prepara un mercato trinomiale con due titoli rischiosi.
}
\examples{
trinomiale(S1 = 1, S2 = 1, r = 0.5, u = c(7/3, 22/9), m = c(1,1), d = c(1/2, 1/3), N = 2)
}
